# PERF: avoid validating indices repeatedly in ArrayManager.reindex_indexer